autotinder/requirements.txt
streamlit==0.73.1
Pillow==7.1.2
numpy==1.19.3
geopy==2.0.0
facenet-pytorch==1.0.1